+++ app/src/main/java/eu/kanade/domain/ui/model/StartScreen.kt
@@ -2,22 +2,19 @@
 package eu.kanade.domain.ui.model
 
 import dev.icerock.moko.resources.StringResource
-import eu.kanade.presentation.util.Tab
-import eu.kanade.tachiyomi.ui.browse.BrowseTab
-import eu.kanade.tachiyomi.ui.library.LibraryTab
-import eu.kanade.tachiyomi.ui.recents.RecentsTab
+import eu.kanade.tachiyomi.ui.home.HomeScreen
 import tachiyomi.i18n.MR
 
-enum class StartScreen(val titleRes: StringResource, val tab: Tab) {
-    ANIME(MR.strings.label_library, LibraryTab),
+enum class StartScreen(val titleRes: StringResource, val tab: HomeScreen.Tab) {
+    ANIME(MR.strings.label_library, HomeScreen.Tab.Library()),
 
     // AM (RECENTS) -->
-    UPDATES(MR.strings.label_recent_updates, RecentsTab),
-    HISTORY(MR.strings.label_recent_manga, RecentsTab),
+    UPDATES(MR.strings.label_recent_updates, HomeScreen.Tab.Recents(toHistory = false)),
+    HISTORY(MR.strings.label_recent_manga, HomeScreen.Tab.Recents(toHistory = true)),
 
     // <-- AM (RECENTS)
     // AM (BROWSE) -->
-    BROWSE(MR.strings.browse, BrowseTab),
+    BROWSE(MR.strings.browse, HomeScreen.Tab.Browse(toExtensions = false)),
     // <-- AM (BROWSE)
 }
 // <-- AY